fix: move neurochemcontext to its own file
