+++ src/app/api/toolkit/route.ts
@@ -1,0 +1,4 @@
+export async function GET() {
+  console.log("I am being logged in toolkit/route.ts");
+  return new Response("Success", { status: 200 });
+}
