+++ src/app/moods/page.tsx
@@ -1,28 +1,16 @@
-'use client'
+"use client";
 
-import { createContext, useState } from 'react'
-import { Cube } from './components/Cube'
-import { SliderBox } from './components/SliderBox'
-
-interface NeurochemState {
-  dopamine: number;
-  serotonin: number;
-  adrenaline: number;
-}
-
-interface NeurochemContextType {
-  neuroState: NeurochemState;
-  setNeuroState: React.Dispatch<React.SetStateAction<NeurochemState>>;
-}
-
-export const NeurochemContext = createContext<NeurochemContextType | null>(null)
+import { useState } from "react";
+import { Cube } from "./components/Cube";
+import { SliderBox } from "./components/SliderBox";
+import { NeurochemContext } from "@/context/NeurochemContext";
 
 export default function MoodsPage() {
   const [neuroState, setNeuroState] = useState({
     dopamine: 5,
     serotonin: 5,
-    adrenaline: 5
-  })
+    adrenaline: 5,
+  });
 
   return (
     <NeurochemContext.Provider value={{ neuroState, setNeuroState }}>
@@ -31,5 +19,5 @@ export default function MoodsPage() {
         <SliderBox />
       </div>
     </NeurochemContext.Provider>
-  )
+  );
 }
